+++ include/boards.h
@@ -7,7 +7,6 @@
 
 #pragma once
 
-#include "settings.h"
 #ifndef UNIT_TEST
 #include <Arduino.h>
 #else
@@ -18,6 +17,8 @@ class Board {
  public:
   String name;
   int versionVoltage;
+  int versionTolerance;
+  int revisionPin;
   int shiftUpPin;
   int shiftDownPin;
   int enablePin;
@@ -27,7 +28,11 @@ class Board {
   int stepperSerialRxPin;
   int auxSerialTxPin;
   int auxSerialRxPin;
+  int ledPin;
   int pwrScaler;
+  float rSense;
+  bool homingSupported;
+  float homingSensitivityScaler;
 };
 
 class Boards {
@@ -41,45 +46,63 @@ class Boards {
 
   Boards() {
 #if defined(SMARTSPIN2K_S3)
-    rev3.name               = r3_NAME;
-    rev3.versionVoltage     = r3_VERSION_VOLTAGE;
-    rev3.shiftUpPin         = r3_SHIFT_UP_PIN;
-    rev3.shiftDownPin       = r3_SHIFT_DOWN_PIN;
-    rev3.enablePin          = r3_ENABLE_PIN;
-    rev3.stepPin            = r3_STEP_PIN;
-    rev3.dirPin             = r3_DIR_PIN;
-    rev3.stepperSerialTxPin = r3_STEPPER_SERIAL_TX;
-    rev3.stepperSerialRxPin = r3_STEPPER_SERIAL_RX;
-    rev3.auxSerialTxPin     = r3_AUX_SERIAL_TX;
-    rev3.auxSerialRxPin     = r3_AUX_SERIAL_RX;
-    rev3.pwrScaler          = r3_PWR_SCALER;
+    rev3.name               = "Revision Three (ESP32-S3)";
+    rev3.versionVoltage     = 1241;
+    rev3.versionTolerance   = 300;
+    rev3.revisionPin        = 4;
+    rev3.shiftUpPin         = 14;
+    rev3.shiftDownPin       = 13;
+    rev3.enablePin          = 48;
+    rev3.stepPin            = 21;
+    rev3.dirPin             = 47;
+    rev3.stepperSerialTxPin = 11;
+    rev3.stepperSerialRxPin = 12;
+    rev3.auxSerialTxPin     = 17;
+    rev3.auxSerialRxPin     = 18;
+    rev3.ledPin             = 2;
+    rev3.pwrScaler          = 12;
+    rev3.rSense             = 0.04f;
+    rev3.homingSupported    = true;
+    rev3.homingSensitivityScaler = 1.6f;
 #else
     // Rev 1
-    rev1.name               = r1_NAME;
-    rev1.versionVoltage     = r1_VERSION_VOLTAGE;
-    rev1.shiftUpPin         = r1_SHIFT_UP_PIN;
-    rev1.shiftDownPin       = r1_SHIFT_DOWN_PIN;
-    rev1.enablePin          = r1_ENABLE_PIN;
-    rev1.stepPin            = r1_STEP_PIN;
-    rev1.dirPin             = r1_DIR_PIN;
-    rev1.stepperSerialTxPin = r1_STEPPER_SERIAL_TX;
-    rev1.stepperSerialRxPin = r1_STEPPER_SERIAL_RX;
+    rev1.name               = "Revision One";
+    rev1.versionVoltage     = 0;
+    rev1.versionTolerance   = 0;
+    rev1.revisionPin        = 34;
+    rev1.shiftUpPin         = 19;
+    rev1.shiftDownPin       = 18;
+    rev1.enablePin          = 13;
+    rev1.stepPin            = 25;
+    rev1.dirPin             = 33;
+    rev1.stepperSerialTxPin = 12;
+    rev1.stepperSerialRxPin = 14;
     rev1.auxSerialTxPin     = 0;
     rev1.auxSerialRxPin     = 0;
-    rev1.pwrScaler          = r1_PWR_SCALER;
+    rev1.ledPin             = 2;
+    rev1.pwrScaler          = 31;
+    rev1.rSense             = 0.08f;
+    rev1.homingSupported    = false;
+    rev1.homingSensitivityScaler = 1.0f;
     // Rev 2
-    rev2.name               = r2_NAME;
-    rev2.versionVoltage     = r2_VERSION_VOLTAGE;
-    rev2.shiftUpPin         = r2_SHIFT_UP_PIN;
-    rev2.shiftDownPin       = r2_SHIFT_DOWN_PIN;
-    rev2.enablePin          = r2_ENABLE_PIN;
-    rev2.stepPin            = r2_STEP_PIN;
-    rev2.dirPin             = r2_DIR_PIN;
-    rev2.stepperSerialTxPin = r2_STEPPER_SERIAL_TX;
-    rev2.stepperSerialRxPin = r2_STEPPER_SERIAL_RX;
-    rev2.auxSerialTxPin     = r2_AUX_SERIAL_TX;
-    rev2.auxSerialRxPin     = r2_AUX_SERIAL_RX;
-    rev2.pwrScaler          = r2_PWR_SCALER;
+    rev2.name               = "Revision Two";
+    rev2.versionVoltage     = 4095;
+    rev2.versionTolerance   = 0;
+    rev2.revisionPin        = 34;
+    rev2.shiftUpPin         = 26;
+    rev2.shiftDownPin       = 32;
+    rev2.enablePin          = 27;
+    rev2.stepPin            = 25;
+    rev2.dirPin             = 33;
+    rev2.stepperSerialTxPin = 19;
+    rev2.stepperSerialRxPin = 18;
+    rev2.auxSerialTxPin     = 21;
+    rev2.auxSerialRxPin     = 22;
+    rev2.ledPin             = 2;
+    rev2.pwrScaler          = 12;
+    rev2.rSense             = 0.08f;
+    rev2.homingSupported    = true;
+    rev2.homingSensitivityScaler = 1.0f;
 #endif
   }
 };